+++ derivatives/__init__.py
@@ -1,3 +1,4 @@
+from .codegen import generate_c, generate_dot
 from .dfa import Dfa, DfaRunner, make_dfa
 from .edsl import (
     Regex, any_char, any_with, any_without, char, char_range, char_set, empty,
@@ -9,4 +10,5 @@ __all__ = [
     "Regex", "Dfa", "DfaRunner", "make_dfa", "any_char", "any_with",
     "any_without", "char", "char_range", "char_set", "empty", "epsilon",
     "string", "tag", "make_lexer", "raise_on_conflict", "select_first",
+    "generate_c", "generate_dot"
 ]
