+++ derivatives/dfa.py
@@ -1,9 +1,6 @@
-import html
 from collections import defaultdict, deque
-from contextlib import contextmanager
-from io import StringIO
 from itertools import count
-from typing import Any, Callable, Dict, Iterator, List, Optional, Set, Tuple
+from typing import Callable, Dict, Iterator, List, Optional, Set, Tuple
 
 from .vector import Vector
 
@@ -48,6 +45,20 @@ class Dfa:
         self._delta = delta
         self._eof_tags = eof_tags
 
+    def iter_delta(self) -> Iterator[Tuple[int, DfaTransitions]]:
+        return enumerate(self._delta)
+
+    def iter_eof_tags(self) -> Iterator[Tuple[int, Optional[str]]]:
+        return enumerate(self._eof_tags)
+
+    def get_tags_set(self) -> Set[str]:
+        tags = {tag for tag in self._eof_tags if tag is not None}
+        for transitions in self._delta:
+            for _, _, tag in transitions:
+                if tag is not None:
+                    tags.add(tag)
+        return tags
+
     def start(self) -> DfaRunner:
         return DfaRunner(self._delta, self._eof_tags)
 
@@ -77,191 +88,6 @@ class Dfa:
             tag, pos = result
             yield tag, input[:pos]
             input = input[pos:]
-
-    def to_dot(self) -> str:
-        def fmt_char(code: int) -> str:
-            char = chr(code)
-            if char in "\\-[]":
-                return "\\" + char
-            return html.escape(char).encode("unicode_escape").decode("ascii")
-
-        buf = Buffer(2)
-        buf.line("digraph dfa {")
-        with buf.indent():
-            buf.line("rankdir=LR")
-            buf.line('"" [shape=none]')
-            buf.line('"" -> "0"')
-            buf.line('"end" [shape=doublecircle]')
-
-            for state, tag in enumerate(self._eof_tags):
-                buf.format('"{}" [shape=circle fixedsize=shape]', state)
-                if tag:
-                    buf.format('"{}" -> "end" [label="EOF/{}"]', state, tag)
-
-            for state, trantisions in enumerate(self._delta):
-                compact: Dict[
-                    Tuple[Optional[int], Optional[str]],
-                    List[Tuple[int, int]]
-                ] = defaultdict(list)
-                last = 0
-                for end, target, tag in trantisions:
-                    compact[(target, tag)].append((last, end - 1))
-                    last = end
-                for (target, tag), ranges in compact.items():
-                    classes = []
-                    for start, end in ranges:
-                        size = end - start + 1
-                        if size == 1:
-                            classes.append(fmt_char(start))
-                        elif size <= 3:
-                            classes.append(
-                                "".join(
-                                    fmt_char(c)
-                                    for c in range(start, end + 1)
-                                )
-                            )
-                        else:
-                            classes.append(
-                                fmt_char(start) + "-" + fmt_char(end)
-                            )
-                    label = "[{}]".format("".join(classes))
-                    if tag is not None:
-                        label += "/" + tag
-                    if target is not None:
-                        buf.format(
-                            '"{}" -> "{}" [label=<{}>]', state, target, label
-                        )
-
-        buf.line("}")
-        return buf.getvalue()
-
-    def to_c(self) -> str:
-        buf = Buffer(4)
-        buf.line("#include <stdint.h>")
-        buf.skip()
-        buf.line("#define DFA_ERROR -1")
-        buf.line("#define DFA_CONTINUE 0")
-        buf.line("#define DFA_MATCH 1")
-        buf.line("#define DFA_END 2")
-        buf.line("#define DFA_END_MATCH 3")
-        buf.skip()
-        tags = {tag for tag in self._eof_tags if tag is not None}
-        for transitions in self._delta:
-            for _, _, tag in transitions:
-                if tag is not None:
-                    tags.add(tag)
-
-        def tag_macro(tag: str) -> str:
-            return "DFA_T_" + tag.upper()
-
-        tokens = sorted(tags)
-        for value, tag in enumerate(tokens):
-            buf.format("#define {} {}", tag_macro(tag), value)
-        buf.skip()
-        buf.line("struct Dfa {")
-        with buf.indent():
-            buf.line("unsigned int state;")
-            buf.line("unsigned int token;")
-        buf.line("};")
-        buf.skip()
-        buf.line("static inline void dfa_reset(struct Dfa *dfa) {")
-        with buf.indent():
-            buf.line("dfa->state = 0;")
-        buf.line("}")
-        buf.skip()
-        buf.line("static const char *dfa_token_name(int t) {")
-        with buf.indent():
-            buf.line("static const char *table[] = {")
-            with buf.indent():
-                for name in tokens:
-                    buf.format('"{}",', name)
-            buf.line("};")
-            buf.format("if (t < 0 || t >= {}) {{ return NULL; }}", len(tokens))
-            buf.format("return table[t];")
-        buf.line("};")
-        buf.skip()
-        buf.line(
-            "static inline int dfa_handle(struct Dfa *dfa, uint32_t c) {"
-        )
-        with buf.indent():
-            buf.line("switch (dfa->state) {")
-            for state, transitions in enumerate(self._delta):
-                buf.format("case {}:", state)
-                with buf.indent():
-                    for end, target, tag in transitions:
-                        if tag is None:
-                            if target is None:
-                                buf.format(
-                                    "if (c < {}) {{ return DFA_END; }}", end
-                                )
-                            else:
-                                buf.format(
-                                    "if (c < {}) {{ dfa->state = {};"
-                                    " return DFA_CONTINUE; }}",
-                                    end, target
-                                )
-                        else:
-                            if target is None:
-                                buf.format(
-                                    "if (c < {}) {{ dfa->token = {};"
-                                    " return DFA_END_MATCH; }}",
-                                    end, tag_macro(tag)
-                                )
-                            else:
-                                buf.format(
-                                    "if (c < {}) {{ dfa->token = {};"
-                                    " dfa->state = {}; return DFA_MATCH; }}",
-                                    end, tag_macro(tag), target
-                                )
-                    buf.line("break;")
-            buf.line("}")
-            buf.line("return DFA_ERROR;")
-        buf.line("}")
-        buf.skip()
-        buf.line("static inline int dfa_handle_eof(struct Dfa *dfa) {")
-        with buf.indent():
-            buf.line("switch (dfa->state) {")
-            for state, tag in enumerate(self._eof_tags):
-                if tag is None:
-                    continue
-                buf.format("case {}:", state)
-                with buf.indent():
-                    buf.format("dfa->token = {};", tag_macro(tag))
-                    buf.format("return DFA_END_MATCH;")
-            buf.line("default:")
-            with buf.indent():
-                buf.format("return DFA_END;")
-            buf.line("}")
-            buf.line("return DFA_ERROR;")
-        buf.line("}")
-        return buf.getvalue()
-
-
-class Buffer:
-    def __init__(self, indent: int = 2):
-        self._buffer = StringIO()
-        self._indent = " " * indent
-        self._level = 0
-
-    @contextmanager
-    def indent(self) -> Iterator[None]:
-        self._level += 1
-        yield
-        self._level -= 1
-
-    def skip(self, n: int = 1) -> None:
-        self._buffer.write("\n" * n)
-
-    def line(self, s: str) -> None:
-        self._buffer.write(self._indent * self._level)
-        self._buffer.write(s)
-        self._buffer.write("\n")
-
-    def format(self, s: str, *args: Any, **kwargs: Any) -> None:
-        self.line(s.format(*args, **kwargs))
-
-    def getvalue(self) -> str:
-        return self._buffer.getvalue()
 
 
 def make_dfa(vector: Vector, tag_resolver: Callable[[Set[int]], str]) -> Dfa:
